wine_region_scraper/requirements.txt
pinecone-client>=6.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
scrapy>=2.11.0
pandas>=2.0.0
lxml>=4.9.0
